kubernetes: Integrate k8s common labels
Towards https://github.com/giantswarm/giantswarm/issues/8632

Include common labels advocated in Kubernetes docs, there's some
overlap with our internal ones and after discussing this the consensus
was that it would be better to avoid that and standardise on the well
defined set from Kubernetes/Helm upstream.

Those labels allow us to describe the source and instance of an app so
it's easy to see which installation an object is part of.

Inline description of the main labels specifying what are the expected
values and that they apply to all objects composing an app/operator
they refer to.

Also add this to description of related custom `.../branch` and
`.../commit` labels.