kubernetes: Integrate k8s common labels (#102)
Towards https://github.com/giantswarm/giantswarm/issues/8632

* Integrate k8s common labels

Include common labels advocated in Kubernetes docs, there's some
overlap with our internal ones and after discussing this the consensus
was that it would be better to avoid that and standardise on the well
defined set from Kubernetes/Helm upstream.

Those labels allow us to describe the source and instance of an app so
it's easy to see which installation an object is part of.

Inline description of the main labels specifying what are the expected
values and that they apply to all objects composing an app/operator
they refer to.

Also add this to description of related custom `.../branch` and
`.../commit` labels.

* Remove `app` and `version` labels

Now replaced with upstream Kubernetes/Helm recommended labels
`app.kubernetes.io/name` and `app.kubernetes.io/version`.

* List labels applied in Helm charts

Add a full list of app labels applied in Helm charts so it's easier
to see without also having to look at external docs.

* Explain maintaining labels using helper template

This allows maintaining them in one place and using in multiple places
without repetition. It is based on approach in default chart template
in upstream Helm.

Also add a collapsed by default example of how this is rendered.